Refactor skill-bonus
